Pass region size as unsigned long long
Even on 32-bit platforms (e.g. x86) there might be physical memory
regions that are larger than 4 GiB (compare with the definition of
l4util_mb_addr_range_t).

To avoid overflow issues and misinterpretation of the size argument
of Region::start_size(), its type needs to be unsigned long long.

+++ server/src/region.h
@@ -93,7 +93,7 @@ public:
    * @param sub the subtype of the region
    * @param eager Mark region as eagerly mapped by kernel
    */
-  static Region start_size(unsigned long long begin, unsigned long size,
+  static Region start_size(unsigned long long begin, unsigned long long size,
                            char const *name = 0, Type t = No_mem,
                            short sub = 0, bool eager = false)
   {
@@ -110,7 +110,7 @@ public:
    * @param sub the subtype of the region
    * @param eager Mark region as eagerly mapped by kernel
    */
-  static Region start_size(void const *begin, unsigned long size,
+  static Region start_size(void const *begin, unsigned long long size,
                            char const *name = 0, Type t = No_mem,
                            short sub = 0, bool eager = false)
   {
